requirements/base.txt
appnope==0.1.0
asgiref==3.2.10
autopep8==1.5.4
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1.2
django-cors-headers==3.5.0
django-extensions==3.0.9
djangorestframework==3.12.1
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.7.2
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
toml==0.10.1
traitlets==5.0.5
urllib3==1.25.11
wcwidth==0.2.5
